fix #11; refactore binding to separate classes
