+++ source/DD4T.DI.Autofac/Bootstrap.cs
@@ -36,66 +36,17 @@ namespace DD4T.DI.Autofac
 
             var load = Assembly.LoadFile(file);
 
-            var provider = AppDomain.CurrentDomain.GetAssemblies().Where(ass => ass.FullName.StartsWith("DD4T.Providers")).FirstOrDefault();
-            var providerTypes = provider.GetTypes();
-            var pageprovider = providerTypes.Where(a => typeof(IPageProvider).IsAssignableFrom(a)).FirstOrDefault();
-            var cpProvider = providerTypes.Where(a => typeof(IComponentPresentationProvider).IsAssignableFrom(a)).FirstOrDefault();
-            var linkProvider = providerTypes.Where(a => typeof(ILinkProvider).IsAssignableFrom(a)).FirstOrDefault();
-            var binaryProvider = providerTypes.Where(a => typeof(IBinaryProvider).IsAssignableFrom(a)).FirstOrDefault();
-            var componentProvider = providerTypes.Where(a => typeof(IComponentProvider).IsAssignableFrom(a)).FirstOrDefault();
-            var commonServices = providerTypes.Where(a => typeof(IProvidersCommonServices).IsAssignableFrom(a)).FirstOrDefault();
+            builder.RegisterProviders();
+            builder.RegisterFactories();
+            builder.RegisterRestProvider();
+            builder.RegisterMvc();
+            builder.RegisterResolvers();
+            builder.RegisterViewModels();
+
 
             builder.RegisterType<DD4TConfiguration>().As<IDD4TConfiguration>().SingleInstance().PreserveExistingDefaults();
-            builder.RegisterType<DefaultPublicationResolver>().As<IPublicationResolver>().SingleInstance().PreserveExistingDefaults();
             builder.RegisterType<DefaultLogger>().As<ILogger>().SingleInstance().PreserveExistingDefaults();
-
             builder.RegisterType<DefaultCacheAgent>().As<ICacheAgent>().PreserveExistingDefaults();
-
-            
-            builder.RegisterRestProviderTypes();
-            builder.RegisterMVCTypes();
-
-            if (commonServices != null)
-                builder.RegisterType(commonServices).As<IProvidersCommonServices>().PreserveExistingDefaults();
-            if (pageprovider != null)
-                builder.RegisterType(pageprovider).As<IPageProvider>().PreserveExistingDefaults();
-            if (cpProvider != null)
-                builder.RegisterType(cpProvider).As<IComponentPresentationProvider>().PreserveExistingDefaults();
-            if (binaryProvider != null)
-                builder.RegisterType(binaryProvider).As<IBinaryProvider>().PreserveExistingDefaults();
-            if (linkProvider != null)
-                builder.RegisterType(linkProvider).As<ILinkProvider>().PreserveExistingDefaults();
-            if (componentProvider != null)
-                builder.RegisterType(componentProvider).As<IComponentProvider>().PreserveExistingDefaults();
-
-
-            builder.RegisterType<FactoryCommonServices>().As<IFactoryCommonServices>().PreserveExistingDefaults();
-            builder.RegisterType<PageFactory>().As<IPageFactory>().PreserveExistingDefaults();
-            builder.RegisterType<ComponentPresentationFactory>().As<IComponentPresentationFactory>().PreserveExistingDefaults();
-            builder.RegisterType<ComponentFactory>().As<IComponentFactory>().PreserveExistingDefaults();
-            builder.RegisterType<BinaryFactory>().As<IBinaryFactory>().PreserveExistingDefaults();
-            builder.RegisterType<LinkFactory>().As<ILinkFactory>().PreserveExistingDefaults();
-
-            //viewmodels
-            builder.RegisterType<ReflectionOptimizer>()
-                 .As<IReflectionHelper>()
-                 .SingleInstance()
-                 .PreserveExistingDefaults();
-
-            builder.RegisterType<DefaultViewModelResolver>()
-               .As<IViewModelResolver>()
-               .SingleInstance()
-               .PreserveExistingDefaults(); 
-
-            builder.RegisterType<ViewModelFactory>()
-               .As<IViewModelFactory>()
-               .SingleInstance()
-               .PreserveExistingDefaults();
-
-            builder.RegisterType<WebConfigViewModelKeyProvider>()
-             .As<IViewModelKeyProvider>()
-             .SingleInstance()
-             .PreserveExistingDefaults();
 
             //caching JMS
             builder.RegisterType<JMSMessageProvider>()
